% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GAAD}
\alias{GAAD}
\title{Type-2 diabetic Gullah-speaking African-Americans dataset}
\format{
A data frame with 288 rows and 7 variables:
}
\usage{
data(GAAD)
}
\description{
A dataset containing periodontal disease status and other attributes of 288 subjects
}
\details{
\itemize{
  \item subPD. the mean periodontal pocket depth (PPD), mm
  \item subCAL. the mean clinical attachment level, mm
  \item Age. age in years
  \item Female. 1 = female, 0 = male
  \item Bmi. body mass index, kg/m^2
  \item Smoker. Smoking status; 1 = smoker, 0 = non-smoker
  \item Hba1cfull. glycated haemoglobin level, %
}
}
\keyword{datasets}
