% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiskewBART.R
\name{Hypers}
\alias{Hypers}
\title{Create hyperparameter object for MultiskewBART}
\usage{
Hypers(
  X,
  Y,
  group = NULL,
  alpha = 1,
  beta = 2,
  gamma = 0.95,
  k = 2,
  num_tree = 20
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nxk matrix of training data response.}

\item{group}{For each column of X, gives the associated group}

\item{alpha}{Positive constant controlling the sparsity level}

\item{beta}{Parameter penalizing tree depth in the branching process prior}

\item{gamma}{Parameter penalizing new nodes in the branching process prior}

\item{k}{Related to the signal-to-noise ratio, sigma_mu = 0.5 / (sqrt(num_tree) * k). BART defaults to k = 2.}

\item{num_tree}{Number of trees in the ensemble}
}
\value{
Returns a list containing the function arguments.
}
\description{
Creates a list which holds all the hyperparameters for use with the MultiskewBART.
}
