% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiskewBART.R
\name{MultiskewBART}
\alias{MultiskewBART}
\title{Fit the MultiskewBART model}
\usage{
MultiskewBART(
  X,
  Y,
  test_X,
  hypers = NULL,
  opts = NULL,
  do_skew = TRUE,
  Wishart = FALSE
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nxk matrix of training data response.}

\item{test_X}{MxP matrix of test data covariates.}

\item{hypers}{a list containing the hyperparameters of the model, usually constructed using the function Hypers().}

\item{opts}{a list containing options for running the chain, usually constructed using the function Opts().}

\item{do_skew}{logical, if true fit the skew-normal model; otherwise, fit a multivariate normal.}

\item{Wishart}{logical, if true use Wishart prior distribution for covariance matrix; otherwise, update the standard deviations and correlations separately.}
}
\value{
Returns a list with the following components:
\itemize{
  \item f_hat_train: fit of the regression function to the training data for
        each iteration of the chain; note that the errors are _not_ mean 0,
        so this does not give the expected value.
  \item f_hat_test: fit of the regression function to the testing data for each
        iteration of the chain; note that the errors are _not_ mean 0, so
        this does not give the expected value.
  \item y_hat_train: predicted values for the training data for
        each iteration of the chain
  \item y_hat_test: predicted values for the testing data for each
        iteration of the chain
  \item Sigma: posterior samples of the covariance matrix
  \item lambda: posterior samples of skewness parameters
  \item loo: the PSIS-loo computed with loo function in the loo package
}
}
\description{
Fits the MultiskewBART model of Um et al. (2021+). The model is of the form \deqn{Y_i = \mu(X_i) + \epsilon_i} where \eqn{\epsilon_i} has a multivariate skew-normal distribution.
}
