% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opts.R
\name{Opts}
\alias{Opts}
\title{MCMC options for MultiskewBART}
\usage{
Opts(
  num_burn = 5000,
  num_thin = 1,
  num_save = 5000,
  num_print = 100,
  update_Sigma_mu = FALSE,
  update_s = FALSE,
  update_alpha = FALSE
)
}
\arguments{
\item{num_burn}{Number of warmup iterations for the chain.}

\item{num_thin}{Thinning interval for the chain.}

\item{num_save}{The number of samples to collect; in total, num_burn + num_save * num_thin iterations are run}

\item{num_print}{Interval for how often to print the chain's progress}

\item{update_Sigma_mu}{If true, Sigma_mu is updated in the Markov chain}

\item{update_s}{If true, s is updated using the Dirichlet prior.}

\item{update_alpha}{If true, alpha is updated using a scaled beta prime prior}
}
\value{
Returns a list containing the function arguments
}
\description{
Creates a list which provides the parameters for running the Markov chain.
}
