% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewBART_parallel.R
\name{skewBART_parallel}
\alias{skewBART_parallel}
\title{Fit the skewBART model with parallel computation}
\usage{
skewBART_parallel(
  X,
  Y,
  test_X,
  hypers = NULL,
  opts = NULL,
  cores = 2L,
  nice = 19L
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nx1 vector of training data response.}

\item{test_X}{MxP matrix of test data covariates}

\item{hypers}{A list of hyperparameters, typically constructed with the UHypers function}

\item{opts}{A list of options for running the Markov chain, typically constructed with the UOpts function.}

\item{cores}{Number of cores to run skewBART in parallel.}

\item{nice}{Job niceness. The niceness scale ranges from 0 (highest) to 19 (lowest);; if not provided, defaults to 19}
}
\value{
Returns a list with the following components:
\itemize{
  \item f_hat_train: posterior samples of function fit to the training data for each iteration of the chain
  \item f_hat_test: posterior samples of function fit to the testing data for each iteration of the chain
  \item y_hat_train: posterior samples of predicted outcome on the training data for each iteration of the chain; not equal to f_hat_train because the errors are not mean 0
  \item y_hat_test: posterior samples of function fit to the testing data for each iteration of the chain; not equal to f_hat_test because the errors are not mean 0
  \item sigma: posterior samples of the scale of the skew-normal distribution
  \item alpha: posterior samples of the skewness parameter of the skew-normal distribution
  \item tau: posterior samples of the standard deviation of the latent normal distribution in the parameter-expanded representation of the skew-normal; usually not of direct interest.
  \item lambda: posterior samples of the regression coefficient of the truncated normal variables in the parameter-expanded representation of the skew-normal distribution; usually not of direct interest.
  \item likelihood_mat: (N x num_save) matrix of log-likelihood values to calculate the log pseudo marginal likelihood (LPML)
}
}
\description{
Fit the skewBART model with parallel computation
}
