% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewBART_pd.R
\name{skewBART_pd}
\alias{skewBART_pd}
\title{Title
Fit skewBART with partial dependence}
\usage{
skewBART_pd(X, Y, vars, x_grid = NULL, hypers = NULL, opts = NULL)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nx1 vector of training data response.}

\item{vars}{the variables to compute the partial dependence function of; must
be a string.}

\item{x_grid}{grid of values along which to compute the partial dependence
function. Must be a data.frame with columns corresponding to the different
columns of X the partial dependence is to be computed on. A default is
provided for continuous variables provided that only one variable is used.}

\item{hypers}{A list of hyperparameters, typically constructed with the
UHypers function}

\item{opts}{A list of options for running the Markov chain, typically
constructed with the UOpts function.}
}
\value{
Returns a list with the following components:
\itemize{
  \item f_hat_train: posterior samples of function fit to the training data for each iteration of the chain
  \item f_hat_test: posterior samples of function fit to the testing data for each iteration of the chain
  \item y_hat_train: posterior samples of predicted outcome on the training data for each iteration of the chain; not equal to f_hat_train because the errors are not mean 0
  \item y_hat_test: posterior samples of function fit to the testing data for each iteration of the chain; not equal to f_hat_test because the errors are not mean 0
  \item sigma: posterior samples of the scale of the skew-normal distribution
  \item alpha: posterior samples of the skewness parameter of the skew-normal distribution
  \item tau: posterior samples of the standard deviation of the latent normal distribution in the parameter-expanded representation of the skew-normal; usually not of direct interest.
  \item lambda: posterior samples of the regression coefficient of the truncated normal variables in the parameter-expanded representation of the skew-normal distribution; usually not of direct interest.
  \item loo: the PSIS-loo computed with loo function in the loo package.
  \item partial_dependence_samples: Samples of the partial dependence function for both y_hat and f_hat. Stored as a data frame with each row corresponding to an iteration of the chain and combination of variables
  \item partial_dependence_summary: Summary of the partial dependence function. Each row corresponds to a combination of variables, and includes the posterior mean and a 95% posterior credible interval for f_hat and y_hat.
}
}
\description{
Fits the skewBART model and returns a list with function estimates and
estimates of the parameters of the skew-normal distribution. Also computes
Friedman's partial dependence function for a specified variable.
See the real_data_GAAD vignette by running browseVignettes("skewBART") for
an illustration.
}
