% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewBART.R
\name{skewBART}
\alias{skewBART}
\title{MCMC options for skewBART}
\usage{
skewBART(
  X = X_norm,
  Y = Y_scaled,
  test_X = test_X_norm,
  tree_num = 20,
  iter = 500,
  burn = 250
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nx1 vector of training data response.}

\item{test_X}{N'xP matrix of test data covariates}

\item{tree_num}{the number of trees}

\item{iter}{Number of MCMC iteration}

\item{burn}{Number of warmup iterations for the chain.}
}
\value{
Returns a list containing the function arguments.
}
\description{
MCMC options for skewBART
}
