% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewBART.R
\name{skewBART}
\alias{skewBART}
\title{MCMC options for skewBART}
\usage{
skewBART(
  X = X_norm,
  Y = Y_scaled,
  test_X = test_X_norm,
  tree_num = 20,
  iter = 500,
  burn = 250
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nx1 vector of training data response.}

\item{test_X}{MxP matrix of test data covariates}

\item{tree_num}{the number of trees}

\item{iter}{Number of MCMC iteration}

\item{burn}{Number of warmup iterations for the chain.}
}
\value{
Returns a list with the following components:
\itemize{
  \item y_hat_train: fit to the training data for each iteration of the chain
  \item y_hat_test: fit to the testing data for each iteration of the chain
  \item sigma: posterior samples of the error standard deviations (sigma^2/(1+alpha^2))
  \item lambda: posterior samples of skewness (alpha*sigma/sqrt(1+alpha^2))
}
}
\description{
MCMC options for skewBART
}
